S 1 2 2

+++ backend/CulturalHeritage/src/main/java/tim2/CulturalHeritage/service/CulturalHeritageServiceImpl.java
@@ -74,6 +74,9 @@ public class CulturalHeritageServiceImpl implements CulturalHeritageService {
 
     @Override
     public void deleteById(Long id) {
+        if(null == culturalHeritageRepository.findById(id).orElse(null))
+            throw new EntityNotFoundException();
+            
         culturalHeritageRepository.deleteById(id);
     }
